+++ lognflow/__init__.py
@@ -1,9 +1,9 @@
 """Top-level package for lognflow."""
 
-__author__ = """Alireza Sadri"""
-__email__ = 'arsadri AT gmail.com'
-__version__ = '0.6.1'
+__author__ = 'Alireza Sadri'
+__email__ = 'arsadri@gmail.com'
+__version__ = '0.6.2'
 
-from .lognflow import lognflow, select_directory, open_file
+from .lognflow import lognflow, select_directory, select_file
 from .logviewer import logviewer
 from .printprogress import printprogress